backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.0
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.13.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.0
pydantic_core==2.23.2
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.38.6
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.30.6
watchfiles==1.1.1
websockets==15.0.1
pymongo>=4.6
passlib[bcrypt]>=1.7
python-jose>=3.3
bcrypt>=4.0.1
